fix : pages folder renamed compeletly
