+++ kadoo/package.json
@@ -31,15 +31,16 @@
     "react-icons": "^4.3.1",
     "react-multi-carousel": "^2.6.5",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3",
+    "react-scripts": "^5.0.1",
     "rechart": "^0.0.1",
     "recharts": "^2.1.6",
+    "sass": "^1.60.0",
     "styled-components": "^5.3.3",
     "swiper": "^6.8.4",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
