+++ kadoo/src/Components/AcceptedTicketCard/AcceptedTicketCard.jsx
