+++ kadoo/src/Components/widgetSm/index.jsx
