+++ kadoo/src/Pages/AddToCart/index.jsx
@@ -1,9 +1,9 @@
-import Header from '../Components/Cart/Header'
-import Main from '../Components/Cart/Main'
-import Basket from '../Components/Cart/Basket'
+import Header from '../../Components/Cart/Header'
+import Main from '../../Components/Cart/Main'
+import Basket from '../../Components/Cart/Basket'
 import React, { useEffect, useState } from 'react'
-import axiosInstance from '../Components/Cart/axios'
-import AppBar from '../Components/AppBar/AppBar'
+import axiosInstance from '../../Components/Cart/axios'
+import AppBar from '../../Components/AppBar/AppBar'
 import 'react-router-dom'
 import { Link } from 'react-router-dom'
 import { styled } from '@mui/material/styles'
