+++ kadoo/src/Pages/AdminHome/index.jsx
@@ -1,8 +1,8 @@
 import Chart from "../../Components/chart/Chart";
 import FeaturedInfo from "../../Components/featuredInfo/FeaturedInfo";
-import "./AdminHome.css";
+import "./style.scss";
 import { userData } from "../../dummyData";
-import WidgetSm from "../../Components/widgetSm/WidgetSm";
+import WidgetSm from "../../Components/widgetSm";
 import WidgetLg from "../../Components/widgetLg/WidgetLg";
 
 export default function AdminHome() {
