+++ kadoo/src/Pages/AdminNewUser/index.jsx
@@ -1,4 +1,4 @@
-import "./AdminNewUser.css";
+import "./style.scss";
 import React from "react";
 import AdapterDateFns from "@mui/lab/AdapterDateFns";
 import LocalizationProvider from "@mui/lab/LocalizationProvider";
