+++ kadoo/src/Pages/AdminPage/index.jsx
@@ -1,15 +1,15 @@
 import Sidebar from '../../Components/sidebar/Sidebar'
 import AppBar from '../../Components/AppBar/AppBar'
-import './AdminPage.css'
-import AdminHome from '../../Pages/AdminHome/AdminHome'
+import './style.scss'
+import AdminHome from '../AdminHome'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
-import AdminUserList from '../../Pages/AdminUserList/AdminUserList'
-import AdminUser from '../../Pages/AdminUser/AdminUser'
-import AdminNewUser from '../../Pages/AdminNewUser/AdminNewUser'
-import AdminProductList from '../../Pages/AdminProductList/AdminProductList'
-import AdminProduct from '../../Pages/AdminProduct/AdminProduct'
-import AdminNewProduct from '../../Pages/AdminNewProduct/AdminNewProduct'
-import UserList from '../../Pages/AdminUserList/AdminUserList'
+import AdminUserList from '../AdminUserList'
+import AdminUser from '../AdminUser'
+import AdminNewUser from '../AdminNewUser'
+import AdminProductList from '../AdminProductList'
+import AdminProduct from '../AdminProduct'
+import AdminNewProduct from '../AdminNewProduct'
+import UserList from '../AdminUserList'
 import { useState, useEffect } from 'react'
 
 function App() {
