+++ kadoo/src/Pages/AdminUserList/index.jsx
@@ -1,4 +1,4 @@
-import "./AdminUserList.css";
+import "./style.scss";
 import { DataGrid } from '@mui/x-data-grid';
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 import { productRows } from "../../dummyData";
