+++ kadoo/src/Pages/HomePage/style.scss
