+++ kadoo/src/Pages/NotFound/index.jsx
@@ -4,7 +4,7 @@ import Box from '@mui/material/Box'
 import Grid from '@mui/material/Grid'
 import Button from '@mui/material/Button'
 import Typography from '@mui/material/Typography'
-import AppBar from '../Components/AppBar/AppBar'
+import AppBar from '../../Components/AppBar/AppBar'
 
 function NotFound() {
   return (
