+++ kadoo/src/Pages/Plantmanagement/style.scss
